Before this PR, the benchmark got a misleading result: `fastest` benchmark was always `20 ns` since no iteration was made with the scheduler, it also returned results in a strange format, and it didn't work as it was since `SUITE` was undefined. Note that to make it work we need to do something in the `step!` since otherwise the compiler tricks the benchmark returning always `20 ns` so I incremented an attribute.

Before simple printing returned:

```julia
13-element BenchmarkTools.BenchmarkGroup:
  tags: ["Schedulers"]
  "large_model" => 11-element BenchmarkTools.BenchmarkGroup:
	  tags: ["by_id", "ByID", "randomly", "Randomly"]
	  "Randomly" => Trial(10.387 ms)
	  "by_id" => Trial(46.445 ms)
	  "ByType" => Trial(57.637 ms)
	  "ByProperty" => Trial(82.430 ms)
	  "fastest" => Trial(20.000 ns) # wrong/not representative
	  "Partially" => Trial(11.445 ms)
	  "by_property" => Trial(70.783 ms)
	  "randomly" => Trial(11.391 ms)
	  "ByID" => Trial(45.150 ms)
	  "by_type" => Trial(135.324 ms)
	  "partially" => Trial(11.310 ms)
  "Randomly" => Trial(2.956 μs)
  "by_id" => Trial(4.248 μs)
  "ByType" => Trial(8.396 μs)
  "ByProperty" => Trial(13.425 μs)
  "fastest" => Trial(20.000 ns) # wrong/not representative
  "small_model" => 11-element BenchmarkTools.BenchmarkGroup:
	  tags: ["by_id", "ByID", "randomly", "Randomly"]
	  "Randomly" => Trial(360.000 ns)
	  "by_id" => Trial(391.000 ns)
	  "ByType" => Trial(922.000 ns)
	  "ByProperty" => Trial(1.332 μs)
	  "fastest" => Trial(20.000 ns) # wrong/not representative
	  "Partially" => Trial(461.000 ns)
	  "by_property" => Trial(1.282 μs)
	  "randomly" => Trial(361.000 ns)
	  "ByID" => Trial(331.000 ns)
	  "by_type" => Trial(4.809 μs)
	  "partially" => Trial(531.000 ns)
  "Partially" => Trial(4.247 μs)
  "by_property" => Trial(12.193 μs)
  "randomly" => Trial(3.106 μs)
  ⋮
```

Now it returns them this way

```julia
3-element BenchmarkTools.BenchmarkGroup:
  tags: ["Schedulers"]
  "large_model" => 11-element BenchmarkTools.BenchmarkGroup:
	  tags: []
	  "Randomly" => Trial(68.026 ms)
	  "by_id" => Trial(86.435 ms)
	  "ByType" => Trial(124.502 ms)
	  "ByProperty" => Trial(109.872 ms)
	  "fastest" => Trial(33.216 ms)
	  "Partially" => Trial(29.824 ms)
	  "by_property" => Trial(91.565 ms)
	  "randomly" => Trial(67.568 ms)
	  "ByID" => Trial(85.417 ms)
	  "by_type" => Trial(329.389 ms)
	  "partially" => Trial(29.280 ms)
  "small_model" => 11-element BenchmarkTools.BenchmarkGroup:
	  tags: []
	  "Randomly" => Trial(651.000 ns)
	  "by_id" => Trial(671.000 ns)
	  "ByType" => Trial(1.272 μs)
	  "ByProperty" => Trial(1.653 μs)
	  "fastest" => Trial(380.000 ns)
	  "Partially" => Trial(632.000 ns)
	  "by_property" => Trial(1.542 μs)
	  "randomly" => Trial(691.000 ns)
	  "ByID" => Trial(601.000 ns)
	  "by_type" => Trial(6.211 μs)
	  "partially" => Trial(731.000 ns)
  "medium_model" => 11-element BenchmarkTools.BenchmarkGroup:
	  tags: []
	  "Randomly" => Trial(33.062 μs)
	  "by_id" => Trial(31.931 μs)
	  "ByType" => Trial(58.501 μs)
	  "ByProperty" => Trial(70.132 μs)
	  "fastest" => Trial(15.470 μs)
	  "Partially" => Trial(28.534 μs)
	  "by_property" => Trial(61.766 μs)
	  "randomly" => Trial(33.333 μs)
	  "ByID" => Trial(30.678 μs)
	  "by_type" => Trial(268.739 μs)
	  "partially" => Trial(29.876 μs)
```

Note that running `step!` as a benchmark is more representative in general than running `schedule` since `schedule` doesn't iterate.